refactor: move jitsi creation to a dedicated and cleaned some type in mediaManager
